refactor: patch critical security and perf issues
This major release addresses critical enterprise-level vulnerabilities, memory leaks, and architectural bottlenecks identified during the security and performance audit.

- cmd/main.go: Fix Dependency Injection for AuthService. Implement a context-aware graceful shutdown with a strict timeout to prevent deadlocks. Initialize the bounded Worker Pool and Rate Limiter cleanup process.
- pkg/media/upload.go: Mitigate Image Decompression Bomb (Pixel Flood) DoS attacks by validating image dimensions via `image.DecodeConfig` before allocating memory.
- pkg/utils/worker.go: Redesign background task execution into a bounded Worker Pool pattern to eliminate Goroutine spikes, preventing CPU/RAM exhaustion under heavy load.
- pkg/utils/pagination.go: Prevent SQL Injection by whitelisting allowed sort columns. Mitigate DoS attacks by enforcing a hard cap on the pagination limit.
- pkg/utils/jwt_utils.go: Resolve runtime type assertion panics by correctly casting the JWT `user_id` payload to `uint`.
- internal/middlewares/auth_middleware.go: Close a critical security bypass by strictly enforcing `TokenVersion` validation across all HTTP methods (including GET). Add safe type assertions to prevent server crashes.
- internal/middlewares/rate_limit_middleware.go: Fix Goroutine memory leaks by replacing the infinite `init()` loop with a context-controlled cleanup worker.
- internal/services/auth.service.go: Adhere strictly to Clean Architecture by migrating email dispatch logic from the handler. Enable proper error logging for background tasks to prevent silent failures.
- internal/services/user_service.go: Implement physical file wiping (`os.Remove`) for user avatars during account purge to prevent orphaned data and disk bloat.
- internal/handlers/*: Remove unused dependencies, strengthen the password policy (min 8 characters), and refactor error responses to return precise underlying error messages for better observability.

+++ pkg/utils/jwt_utils.go
@@ -13,10 +13,10 @@ func GetUserIDFromContext(c *gin.Context) (uint, error) {
 		return 0, errors.New("không tìm thấy thông tin xác thực")
 	}
 
-	userIDVal, ok := userIDFloat.(float64)
+	userIDVal, ok := userIDFloat.(uint)
 	if !ok {
 		return 0, errors.New("token sai định dạng")
 	}
 
-	return uint(userIDVal), nil
+	return userIDVal, nil
 }
